+++ tests/test_config_manager.py
@@ -392,6 +392,47 @@ class PollingConfigManagerTest(base.BaseTest):
         )
         self.assertEqual(test_headers['Last-Modified'], project_config_manager.last_modified)
         self.assertIsInstance(project_config_manager.get_config(), project_config.ProjectConfig)
+        self.assertTrue(project_config_manager.is_running)
+
+    def test_fetch_datafile__exception_raised(self, _):
+        """ Test that config_manager keeps running if exception is raised when fetching datafile. """
+        class MockExceptionResponse(object):
+            def raise_for_status(self):
+                raise requests.exceptions.RequestException('Error Error !!')
+
+        sdk_key = 'some_key'
+        mock_logger = mock.Mock()
+        with mock.patch('optimizely.config_manager.PollingConfigManager.fetch_datafile'):
+            project_config_manager = config_manager.PollingConfigManager(sdk_key=sdk_key, logger=mock_logger)
+        expected_datafile_url = enums.ConfigManager.DATAFILE_URL_TEMPLATE.format(sdk_key=sdk_key)
+        test_headers = {'Last-Modified': 'New Time'}
+        test_datafile = json.dumps(self.config_dict_with_features)
+        test_response = requests.Response()
+        test_response.status_code = 200
+        test_response.headers = test_headers
+        test_response._content = test_datafile
+        with mock.patch('requests.get', return_value=test_response):
+            project_config_manager.fetch_datafile()
+
+        self.assertEqual(test_headers['Last-Modified'], project_config_manager.last_modified)
+        self.assertIsInstance(project_config_manager.get_config(), project_config.ProjectConfig)
+
+        # Call fetch_datafile again, but raise exception this time
+        with mock.patch('requests.get', return_value=MockExceptionResponse()) as mock_requests:
+            project_config_manager.fetch_datafile()
+
+        mock_requests.assert_called_once_with(
+            expected_datafile_url,
+            headers={'If-Modified-Since': test_headers['Last-Modified']},
+            timeout=enums.ConfigManager.REQUEST_TIMEOUT,
+        )
+        mock_logger.error.assert_called_once_with('Fetching datafile from {} failed. Error: Error Error !!'.format(
+            expected_datafile_url
+        ))
+        self.assertEqual(test_headers['Last-Modified'], project_config_manager.last_modified)
+        self.assertIsInstance(project_config_manager.get_config(), project_config.ProjectConfig)
+        # Confirm that config manager keeps running
+        self.assertTrue(project_config_manager.is_running)
 
     def test_is_running(self, _):
         """ Test that polling thread is running after instance of PollingConfigManager is created. """
@@ -402,34 +443,34 @@ class PollingConfigManagerTest(base.BaseTest):
 
 @mock.patch('requests.get')
 class AuthDatafilePollingConfigManagerTest(base.BaseTest):
-    def test_init__access_token_none__fails(self, _):
-        """ Test that initialization fails if access_token is None. """
+    def test_init__datafile_access_token_none__fails(self, _):
+        """ Test that initialization fails if datafile_access_token is None. """
         self.assertRaisesRegexp(
             optimizely_exceptions.InvalidInputException,
-            'access_token cannot be empty or None.',
+            'datafile_access_token cannot be empty or None.',
             config_manager.AuthDatafilePollingConfigManager,
-            access_token=None
+            datafile_access_token=None
         )
 
-    def test_set_access_token(self, _):
-        """ Test that access_token is properly set as instance variable. """
-        access_token = 'some_token'
+    def test_set_datafile_access_token(self, _):
+        """ Test that datafile_access_token is properly set as instance variable. """
+        datafile_access_token = 'some_token'
         sdk_key = 'some_key'
         with mock.patch('optimizely.config_manager.AuthDatafilePollingConfigManager.fetch_datafile'):
             project_config_manager = config_manager.AuthDatafilePollingConfigManager(
-                access_token=access_token, sdk_key=sdk_key)
+                datafile_access_token=datafile_access_token, sdk_key=sdk_key)
 
-        self.assertEqual(access_token, project_config_manager.access_token)
+        self.assertEqual(datafile_access_token, project_config_manager.datafile_access_token)
 
     def test_fetch_datafile(self, _):
         """ Test that fetch_datafile sets authorization header in request header and sets config based on response. """
-        access_token = 'some_token'
+        datafile_access_token = 'some_token'
         sdk_key = 'some_key'
         with mock.patch('optimizely.config_manager.AuthDatafilePollingConfigManager.fetch_datafile'), mock.patch(
             'optimizely.config_manager.AuthDatafilePollingConfigManager._run'
         ):
             project_config_manager = config_manager.AuthDatafilePollingConfigManager(
-                access_token=access_token, sdk_key=sdk_key)
+                datafile_access_token=datafile_access_token, sdk_key=sdk_key)
         expected_datafile_url = enums.ConfigManager.AUTHENTICATED_DATAFILE_URL_TEMPLATE.format(sdk_key=sdk_key)
         test_headers = {'Last-Modified': 'New Time'}
         test_datafile = json.dumps(self.config_dict_with_features)
@@ -445,7 +486,8 @@ class AuthDatafilePollingConfigManagerTest(base.BaseTest):
 
         mock_request.assert_called_once_with(
             expected_datafile_url,
-            headers={'Authorization': 'Bearer {access_token}'.format(access_token=access_token)},
+            headers={'Authorization': 'Bearer {datafile_access_token}'.format(
+                datafile_access_token=datafile_access_token)},
             timeout=enums.ConfigManager.REQUEST_TIMEOUT,
         )